Refactor fix_keys utility and remove duplication
Moved the fix_keys function and related regex definitions from mens.py and womens.py into utils.py to eliminate code duplication. Updated imports in mens.py and womens.py to use the shared fix_keys from utils.py. No functional changes to logic; this is a code cleanup for maintainability. Also, minor test data adjustment in test_federation.py.
