+++ ArWikiCats/new_resolvers/jobs_resolvers/mens.py
@@ -8,7 +8,7 @@ from ...helps import logger
 from ...translations import all_country_with_nat_ar, jobs_mens_data, RELIGIOUS_KEYS_PP, all_country_with_nat
 from ...translations_formats import format_multi_data_v2, MultiDataFormatterBaseV2
 from ..nats_as_country_names import nats_keys_as_country_names, nats_keys_as_country_names_bad_keys
-from .utils import one_Keys_more_2, nat_and_gender_keys
+from .utils import fix_keys, one_Keys_more_2, nat_and_gender_keys
 
 countries_en_keys = [x.get("en") for x in all_country_with_nat.values() if x.get("en")]
 
@@ -218,21 +218,6 @@ def load_bot() -> MultiDataFormatterBaseV2:
         use_other_formatted_data=True,
         search_first_part=True,
     )
-
-
-def fix_keys(category: str) -> str:
-    category = category.replace("'", "").lower()
-    category = REGEX_THE.sub("", category)
-    category = re.sub(r"\s+", " ", category)
-    replacements = {
-        "expatriates": "expatriate",
-        "canadian football": "canadian-football",
-    }
-
-    for old, new in replacements.items():
-        category = category.replace(old, new)
-
-    return category.strip()
 
 
 @functools.lru_cache(maxsize=10000)
